dm8000: Fix the /boot problems, don't mount /boot
This should solve the box going into "stop" mode after an upgrade of
the kernel.

Always mount /boot with the compr=none option so that the bootloader
can read it. Compressing /boot is pointless anyway, it only contains
compressed files and a few symlinks.

/boot was being mounted using a "hack" to edit /etc/fstab. Since the
box has no need for /boot to be mounted anyway, don't mount it but
only do that when needed, thus saving a bit of RAM too.